extracted type definitions to a seperate file
